requirements.txt
absl-py==1.4.0
accelerate==0.25.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.8
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asteroid==0.6.0
asteroid-filterbanks==0.4.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
audioldm==0.0.21
audioread==2.1.9
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bitarray==2.8.2
bitsandbytes==0.37.2
black==23.3.0
blessed==1.20.0
blinker==1.6.2
blis==0.7.9
braceexpand==0.1.7
Brotli==1.0.9
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.3.0
catalogue==2.0.8
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==2.0.12
click==8.1.3
cmake==3.27.0
colorama==0.4.6
confection==0.0.4
contexttimer==0.3.3
contourpy==1.0.7
croniter==1.4.1
cryptography @ file:///croot/cryptography_1673298753778/work
cycler==0.11.0
cymem==2.0.7
Cython==3.0.0
dataclasses==0.6
datasets==2.15.0
dateutils==0.6.12
decorator==5.1.1
decord==0.6.0
deepdiff==6.3.1
deepspeed==0.9.0
dill==0.3.6
distlib==0.3.6
docopt==0.6.2
editdistance==0.6.2
einops==0.6.0
-e git+https://github.com/facebookresearch/encodec.git@0e2d0aed29362c8e8f52494baf3e6f99056b214f#egg=encodec
entrypoints==0.4
evaluate==0.4.0
executing==1.2.0
fairscale==0.4.4
fairseq==0.12.2
fast-bss-eval==0.1.4
fastapi==0.94.1
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.12.0
fire==0.5.0
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.39.0
frozenlist==1.3.3
fschat==0.2.33
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gradio==3.21.0
grpcio==1.51.1
h11==0.14.0
hjson==3.1.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.19.4
hydra-core==1.0.7
identify==2.5.23
idna @ file:///croot/idna_1666125576474/work
imageio==2.28.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflate64==0.3.1
inquirer==3.1.3
iopath==0.1.10
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jiwer==3.0.2
joblib==1.1.0
jsonargparse==4.22.1
jsonschema==4.17.3
julius==0.2.7
kaggle==1.5.13
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
librosa==0.9.2
lightning-cloud==0.5.37
lightning-utilities==0.9.0
linkify-it-py==1.0.3
lit==16.0.6
llvmlite==0.38.0
loralib==0.1.1
lxml==4.9.3
Markdown==3.4.1
markdown-it-py==2.1.0
markdown2==2.4.11
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mir-eval==0.7
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mlp-mixer-pytorch==0.1.1
more-itertools==9.1.0
multidict==6.0.4
multiprocess==0.70.14
multivolumefile==0.2.3
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nh3==0.2.14
ninja==1.11.1
nlp-tools @ git+https://github.com/yuyusica/NLP_tools.git@f1e34676d5a82b28a09c3194dd6357aa33c3d434
nltk==3.8.1
nodeenv==1.7.0
num2words==0.5.12
numba==0.55.1
numpy==1.21.1
oauthlib==3.2.2
omegaconf==2.0.6
openai==0.27.6
openai-whisper @ git+https://github.com/openai/whisper.git@0a60fcaa9b86748389a656aa013c416030287d47
opencv-python==4.7.0.68
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.8.7
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
pathspec==0.11.1
pathy==0.10.1
pb-bss-eval==0.0.2
peft @ git+https://github.com/huggingface/peft.git@e536616888d51b453ed354a6f1e243fecb02ea08
pesq==0.0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
plotly==5.14.1
pooch==1.6.0
portalocker==2.7.0
pre-commit==3.2.2
preshed==3.0.8
progressbar==2.5
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py7zr==0.20.5
pyarrow==11.0.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.1
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.18.0
pydantic==1.10.6
pydeck==0.8.1b0
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.14.0
PyJWT==2.8.0
Pympler==1.0.1
pynini==2.1.5
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.8
pyppmd==1.0.0
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pystoi==0.3.3
python-dateutil==2.8.2
python-editor==1.0.4
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.1
pytorch-crf==0.7.2
pytorch-lightning==1.9.1
pytorch-ranger==0.1.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzstd==0.15.9
rapidfuzz==2.13.7
readchar==4.0.5
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
resampy==0.2.2
responses==0.18.0
rfc3986==1.5.0
rich==13.3.1
rouge-score==0.1.2
rsa==4.9
sacrebleu==2.3.1
safetensors==0.4.1
scikit-image==0.20.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.13.0
sentencepiece==0.1.98
seqeval==1.2.2
shortuuid==1.0.11
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
-e git+https://github.com/ZhangXInFD/SpeechTokenizer.git@adc2c3fc65a2eb82efb4744b66905ed018d5895a#egg=speechtokenizer
srsly==2.4.6
stack-data==0.6.2
starlette==0.26.1
starsessions==1.3.0
streamlit==1.21.0
svgwrite==1.4.3
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
text-unidecode==1.3
texttable==1.6.7
thinc==8.1.9
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.3.3
timm==0.4.12
tn==0.0.4
tokenize-rt==5.0.0
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torch-mir-eval==0.4
torch-optimizer==0.1.0
torch-stoi==0.1.2
torchaudio==0.13.1
torchlibrosa==0.0.9
torchmetrics==0.7.3
torchvision==0.14.1
tornado==6.3.1
tqdm==4.64.1
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers.git@e6dcf8abd6f65bb4b6dfc1831b20d9ba49ce00e2
triton==2.0.0
typer==0.7.0
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
tzdata==2023.3
tzlocal==4.3
uc-micro-py==1.0.1
urllib3 @ file:///croot/urllib3_1673575502006/work
uvicorn==0.21.0
validators==0.20.0
virtualenv==20.22.0
wasabi==1.1.1
watchdog==3.0.0
wavedrom==2.0.3.post3
wcwidth==0.2.6
webdataset==0.2.48
websocket-client==1.6.1
websockets==10.4
Werkzeug==2.2.3
WeTextProcessing==0.1.2
-e git+https://github.com/patrickvonplaten/whisper.git@d1690a4d0e0802dcc05ff562f89b4ca65324ceee#egg=whisper
whisper-normalizer==0.0.2
xxhash==3.2.0
yarl==1.8.2
zipp==3.13.0
